requirements.txt
# Python==3.10.11
# pip==23.0.1

# python -m venv env
# env\Scripts\activate

# pip install -r requirements.txt

discord==2.3.2
python-dotenv==1.0.1